+++ src/components/common/Tabs/styles.js
