+++ src/components/common/index.js
@@ -5,3 +5,7 @@ export { Input } from './Input';
 export { SEO } from './SEO';
 export { Divider } from './Divider';
 export { SectionHeader } from './SectionHeader';
+export { Header } from './Header';
+export { Footer } from './Footer';
+export { Tabs } from './Tabs';
+export { Tab } from './Tab';
