--- src/components/theme/index.js
@@ -1,4 +1,0 @@
-export { Header } from './Header';
-export { Footer } from './Footer';
-export { Tabs } from './Tabs';
-export { Tab } from './Tab';
